app/requirements.txt
streamlit
tensorflow==2.15.0
keras==2.15.0
numpy
pillow
opencv-python
scipy
matplotlib
